pyproject.toml
[project]
name = "lambda_clock"
version = "0.1.0"
description = ""
authors = [
    {name = "Tony Holroyd",email = "testingtony12@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
]

[tool.poetry]
packages = [{include = "clocklambda", from = "src"}]


[tool.poetry.group.dev.dependencies]
boto3 = "^1.37.23"
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
pillow = ">=11.2.1,<12.0.0"
pre-commit = "^4.2.0"
ruff = "^0.11.4"

[tool.poetry.group.local.dependencies]
boto3 = "^1.37.27"

[tool.pytest.ini_options]
addopts = "--doctest-modules"

[tool.ruff]
line-length = 120
indent-width = 4
fix = true
exclude = [
    "tests"
]

[tool.ruff.lint]
extend-select = ["E501", "B", "D", "SIM", "PTH", "I", "F", "E", "C90", "N", "UP", "A", "DTZ", "Q"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
